Fix stocks_minute schema and update pipeline config
- Change transactions field from uint16 to uint32 to handle high transaction counts
- Update data_root to use /Volumes/sandisk path
- Disable binary conversion (Parquet-only storage for pandas/DuckDB)
- Update test to match schema changes

+++ src/storage/schemas.py
@@ -66,7 +66,7 @@ def get_stocks_minute_schema() -> pa.Schema:
         pa.field('low', pa.float32()),
         pa.field('close', pa.float32()),
         pa.field('volume', pa.uint32()),
-        pa.field('transactions', pa.uint16()),
+        pa.field('transactions', pa.uint32()),
 
         # Enriched features
         pa.field('alpha_minute', pa.float32(), nullable=True),
